split task_01 to .hpp .cpp files
